Move determinate prompting to separate module, consolidate functionality
